+++ sing-box/route/rule/rule_network_interface_address.go
@@ -40,11 +40,11 @@ func NewNetworkInterfaceAddressItem(networkManager adapter.NetworkManager, inter
 
 func (r *NetworkInterfaceAddressItem) Match(metadata *adapter.InboundContext) bool {
 	interfaces := r.networkManager.NetworkInterfaces()
-	myInterface := r.networkManager.InterfaceMonitor().MyInterface()
+	myInterfaces := r.networkManager.InterfaceMonitor().MyInterfaces()
 match:
 	for ifType, addresses := range r.interfaceAddresses {
 		for _, networkInterface := range interfaces {
-			if networkInterface.Name == myInterface {
+			if common.Contains(myInterfaces, networkInterface.Name) {
 				continue
 			}
 			if networkInterface.Type != ifType {