+++ sing-box/docs/changelog.md
@@ -2,7 +2,11 @@
 icon: material/alert-decagram
 ---
 
-#### 1.14.0-alpha.28
+#### 1.14.0-alpha.29
+
+* Fixes and improvements
+
+#### 1.13.13
 
 * Fixes and improvements